fixed a bug where i was stupid enough to put a 0 in memory where i should have :"""""D

+++ Crow/Engine/Core/ResourceManager.cpp
@@ -201,7 +201,7 @@ Sprite *ResourceManager::GetSprite(const std::string &name)
     return iterator->second;
 }
 
-void ResourceManager::LoadAssetFromAssetsFile(const std::string& filename,int maxAssets)
+void ResourceManager::LoadAssetFromAssetsFile(const std::string& filename,size_t maxAssets)
 {
     AssetCollection assetCollection = ParseAssetFile((ASSET_PATH + filename).data(),maxAssets);
